// client/src/types/Patient.ts
export type Patient = {
  _id: string;
  firstName: string;
  lastName: string;
  patientID: string;
  streetAddress: string;
  city: string;
  state: string;
  country: string;
  countryCode: string;
  phoneNumber: string;
  email: string;
  medicalStatus: string;
  createdAt: Date;
  updatedAt: Date;
};